+++ CMakeLists.txt
@@ -12,7 +12,9 @@ include_directories(include)
 
 add_executable(main src/main.cpp src/FPS.cpp src/LTexture.cpp src/LTimer.cpp src/Square.cpp
         src/SceneStateMachine.cpp
-        src/SceneTest.cpp)
+        src/SceneTest.cpp
+        src/Game.cpp
+        src/GameObject.cpp)
 
 target_compile_features(main PRIVATE cxx_std_20)
